mtk: hostapd: add support for DFS channel switching with csa sent
Add support for DFS channel switch

When searching for a DFS channel for a background radar channel switch,
we should not base the selection on the cap of the original channel.
Otherwise, the selected channel may become invalid when the bandwidth is changed.
For example, iface->conf->secondary_channel is set to 0 when operating o BW 20.
Therefore, if the user tries to switch from BW 20 to BW 80, the pre-selected channel
for BW 80 will not be checked by dfs_is_chan_allowed in dfs_find_channel.

Allow user setting the CSA count for DFS channel switch when background
radar is disabled.

Clear all the CS params when the switch to the DFS channel is aborted.
Otherwise, the csa after beacon might contain the old CSA IE.
This happens especially when radar is detected during the CAC period of
the target DFS channel.

Refactor the patch for supporting bandwidth/expand ch configuration in
set_offchain command.
1. Set the default value of new_chwidth to -1 to allow switching to bw 20
   (new_chwidth = CONF_OPER_CHWIDTH_USE_HT = 0)
2. Use the center freq instead of control freq to calculate the expanded
   channel.

The default value of new_chwidth has now been changed to -1, so
the condition should be new_chwidth < 0 instead of !new_chwidth.

Allow channel switch during post-CSA radar detection.
Revert the following upstream patch
"Add support to switch channel when CAC is in progress"

Remove the declaration of dfs_adjust_center_freq

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>