mtk: hostapd: add support for enable/disable single BSS
This function is ported from wifi 6.
(https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/8407972)

Enabling or disabling single BSS mean that the beacon of the BSS is
enabled or disabled.

When the BSS is disabled, the following statements are true:
1. the events or packets from the driver are all ignored.
2. Per radio functions (ex. channel switch, color change) will not
   execute for the BSS, but are finished by other enabling BSS(es).
   The changed parameters take effect when the BSS is enabled again.
3. Enabling the BSS will not reload the configuration. In other word,
   if the configuration changes during the BSS disabling, the BSS needs
   to be removed and added again, not just be enabled.

This patch add new commands to enable/disable single BSS:
$ hostapd_cli -i <ifname> enable_bss
$ hostapd_cli -i <ifname> disable_bss

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>